refactor: rename the file of the `POVMPubResult`
