+++ test/sampler/test_povm_pub_result.py
